Tmux -CC: disable win32 input mode in the control pane (#6768)
On windows, if win32 input mode is active, key presses are encoded by the GUI and will bypass
the key_down logic that we use to detect when the user presses `q` to exit from `tmux -CC` mode.

This commit forces the keyboard encoding to be standard Xterm encoding for the pane
that is running `tmux -CC` mode so that we can catch the `q` press.