Add alphabet to launcher args (#6677)
* feat: add alphabet argument to ShowLauncherArgs

- The labels are now based on the provided argument or the default value
- The 'alphabet' string defaults to "1234567890abcdefghilmnopqrstuvwxyz"
- Optional argument 'alphabet' is added to 'ShowLauncherArgs'

* docs: update documentation for ShowLauncherArgs

* fix: push (max_label_len+3) whitespaces if always_fuzzy false in InputSelector

This is to handle the edge case in InputSelector when max_label_len is 2
and the labels shift one place to the left when entering fuzzy mode and
back one place to the right when exiting fuzzy mode.

* fix: push (max_label_len+3) whitespaces if 'FUZZY' not present

This is to handle the edge case in Launcher when max_label_len is 2
and the labels shift one place to the left when entering fuzzy mode and
back one place to the right when exiting fuzzy mode.

* feat: add configuration option for Launcher alphabet

- This is in order to be able to configure the alphabet for
  actions corresponding to KeyAssignment enums 'ShowLauncher',
  'ShowTabNavigator'
- If `launcher_alphabet` configuration option not set, the value
  defaults to '1234567890abcdefghilmnopqrstuvwxyz'

* docs: add launcher_alphabet documentation

* docs: update documentation for ShowLauncherArgs

Co-authored-by: Wez Furlong <wez@wezfurlong.org>