termwiz: add is_white_space_(char|grapheme) helper functions
These are not really expected to be used by most applications.
They are being added to the cell module to support some higher
level checks as part of trying to handle these wezterm issues
a bit better:

 https://github.com/wezterm/wezterm/issues/1422
 https://github.com/wezterm/wezterm/issues/6637

These functions use the White_Space properties as defined by
the UCD version 16.

I'm using the trie form of the data here because I'm already
using the ucd-trie module on other properties.  The dataset
for White_Space is quite small, and is probably fine to use
a binary search on the non-ucd form of the data, I just didn't
have an example of that code hanging around and didn't want
to have to get distracted debugging that; I somehow always
end up with an off-by-one somewhere when hooking up a binary
search for the first time!

This commit has no functional impact on termwiz or any
other crate in the wezterm repo.