feat(mark-last-tab): append last tab name with '*' (#6895)
* feat(mark-last-tab): append last tab name with '*'

I think there should be a better method instead of make the save function so crowded.
But this works for now similar to `tmux` windows. But, `tmux` is using '*' for current and '-' for last
Hardcoded '*'

* Revert "feat(mark-last-tab): append last tab name with '*'"

This reverts commit 6f11c16dd9b034977088820fbc86a4456df9d513.

* feat(tab): add is_last to indicate previous active tab - last tab.

Add usage example to format-tab-title.md
Add docs to TabInformation.md

Couldn't add tests, perhapse can you guide on adding test for this?

* Update docs/config/lua/TabInformation.md

renaming as suggested.

Co-authored-by: Wez Furlong <wez@wezfurlong.org>

* Update wezterm-gui/src/termwindow/mod.rs

renaming as suggested.

Co-authored-by: Wez Furlong <wez@wezfurlong.org>

* Update docs/config/lua/window-events/format-tab-title.md

Co-authored-by: Wez Furlong <wez@wezfurlong.org>

* refactor: rename `TabInformation` attribute to `is_last_active`

---------

Co-authored-by: Omar Elmasri <omarmasri@duck.com>
Co-authored-by: Wez Furlong <wez@wezfurlong.org>