ssh: implement ServerAliveInterval for libssh backend
This is a partial implementation of keepalives; it has the good
part (the sending of the packet on the requested interval), but
not the other part (disconnecting if we don't hear anything after
some number of packets).

The reason for not having the latter is that it isn't obvious to me how
to expose whether the channel had more incoming packets since sending
the ignore, so we cannot confidently act to terminate the connection.

refs: https://github.com/wezterm/wezterm/issues/4023